requirements.txt
streamlit
backports.lzma
azure-cognitiveservices-vision-customvision
azure-cognitiveservices-vision-face
Pillow
